fixes #87: admin can now delete edit comments

+++ grails-app/services/au/org/ala/biocollect/CommentService.groovy
@@ -16,17 +16,25 @@ class CommentService {
     }
 
     def deleteComment(data){
-        webService.deleteWrapper(grailsApplication.config.ecodata.service.url + "/comment/${data['id']}?userId=${data['userId']}")
+        Boolean alaAdmin = authService.userInRole(grailsApplication.config.security.cas.adminRole)
+        webService.deleteWrapper(grailsApplication.config.ecodata.service.url + "/comment/${data['id']}?userId=${data['userId']}&isALAAdmin=${alaAdmin}")
     }
 
     def updateComment(data){
+        data['isALAAdmin'] = authService.userInRole(grailsApplication.config.security.cas.adminRole)
         webService.doPost(grailsApplication.config.ecodata.service.url + "/comment/${data['id']}",data)
     }
 
     def listComments(data) {
+        String userId = authService.getUserId()
+        Boolean admin = authService.userInRole(grailsApplication.config.security.cas.adminRole)
         def response = webService.doGet(grailsApplication.config.ecodata.service.url + "/comment", data)
         if(response?.resp){
             response.resp['userId'] = authService.getUserId()
+            Map privilege = webService.doGet(grailsApplication.config.ecodata.service.url + "/comment/doesUserHavePrivilege",
+                [userId:response.resp['userId'], entityId:data['entityId'], entityType: data['entityType']] ).resp;
+
+            response.resp['admin'] = privilege?.isAdmin || admin
         }
         response
     }
